io-stats: Change latency to nanoseconds from microseconds (#1833)
- In 'BUMP_THROUGHPUT' macro changed 'elapsed' from microseconds to nanoseconds.
- In 'update_ios_latency' function 'elapsed' is now in nanoseconds.
- In 'collect_ios_latency_sample' function removed conversion from nano to micro,
  instead directly assigned as 'nano' to 'tv_nsec' of 'timespec' macro
- in 'ios_sample_t' macro changed 'timeval' to 'timespec' to support above change.
- In '_io_stats_write_latency_sample' function changed formula to from 1e+6 to 1e+9
  since 'ios_sample_t' macro now has 'timespec'
- In 'BUMP_THROUGHPUT','_ios_sample_t','collect_ios_latency_sample' & update_ios_latency'
  changed 'elapsed' datatype from 'double' to 'int64_t'
- In glusterfs/libglusterfs/src/glusterfs/common-utils.h changed return type of
  'gf_tsdiff' function from 'double' to 'int64_t' since it can return negative values.
- In glusterfs/libglusterfs/src/latency.c, libglusterfs/src/glusterfs/common-utils.h,
  xlators/debug/io-stats/src/io-stats.c & xlators/storage/posix/src/posix-helpers.c
  'elapsed' is now of type 'int64_t'

Fixes: #1825

+++ libglusterfs/src/latency.c
@@ -41,7 +41,7 @@ gf_latency_update(gf_latency_t *lat, struct timespec *begin,
         return;
     }
 
-    double elapsed = gf_tsdiff(begin, end);
+    int64_t elapsed = gf_tsdiff(begin, end);
 
     if (lat->max < elapsed)
         lat->max = elapsed;
